enable the user to add himself to several weddings

+++ src/main/java/com/weeding/time/app/dto/WeddingDto.java
@@ -19,5 +19,5 @@ public class WeddingDto {
     private String location;
     private String accessCode;
     private LocalDateTime createdAt;
-    private Long applicationUserId;
+    private Integer applicationUserId;
 }
